+++ CHANGELOG.md
@@ -5,14 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
-
+## 1.5.0 - 2019-10-07
 ### Added
 - Add new process to suspend/resume ASG process `AlarmNotification` while aem_offline_snapshot lambda is running [shinesolutions/aem-aws-stack-builder#295]
 
 ### Fixed
 - Fix print of wrong instance count while updating publish-dispatcher ASG.
-
 
 ## 1.4.0 - 2019-09-17
 ### Added